requirements.txt
joblib
flask
scikit-learn
flask_sqlalchemy